examples/live/requirements.txt
pyalsaaudio==0.11.0
../../
